+++ dhcp_test.go
@@ -1,6 +1,19 @@
 package dhcp
 
-/*
+import (
+	"context"
+	"net"
+	"net/netip"
+	"testing"
+
+	"github.com/go-logr/logr"
+	"github.com/insomniacslk/dhcp/dhcpv4"
+	"github.com/insomniacslk/dhcp/dhcpv4/nclient4"
+	"github.com/tinkerbell/dhcp/data"
+	"golang.org/x/net/ipv4"
+	"golang.org/x/net/nettest"
+)
+
 type mock struct {
 	Log         logr.Logger
 	ServerIP    net.IP
@@ -11,7 +24,7 @@ type mock struct {
 	Router      net.IP
 }
 
-func (m *mock) Handle(ctx context.Context, conn net.PacketConn, d data.Packet) {
+func (m *mock) Handle(ctx context.Context, conn *ipv4.PacketConn, d data.Packet) {
 	if m.Log.GetSink() == nil {
 		m.Log = logr.Discard()
 	}
@@ -33,7 +46,8 @@ func (m *mock) Handle(ctx context.Context, conn net.PacketConn, d data.Packet) {
 		m.Log.Error(err, "error creating reply")
 		return
 	}
-	if _, err := conn.WriteTo(reply.ToBytes(), d.Peer); err != nil {
+	cm := &ipv4.ControlMessage{IfIndex: d.Md.IfIndex}
+	if _, err := conn.WriteTo(reply.ToBytes(), cm, d.Peer); err != nil {
 		m.Log.Error(err, "failed to send reply")
 		return
 	}
@@ -54,8 +68,6 @@ func (m *mock) setOpts() []dhcpv4.Modifier {
 	return mods
 }
 
-
-
 func dhcp(ctx context.Context) (*dhcpv4.DHCPv4, error) {
 	rifs, err := nettest.RoutedInterface("ip", net.FlagUp|net.FlagBroadcast)
 	if err != nil {
@@ -73,24 +85,23 @@ func dhcp(ctx context.Context) (*dhcpv4.DHCPv4, error) {
 	return c.DiscoverOffer(ctx)
 }
 
-
-func TestListenAndServe(t *testing.T) {
-	// test if the server is listening on the correct address and port
+func TestServe(t *testing.T) {
 	tests := map[string]struct {
-		h            Handler
-		addr         netip.AddrPort
-		wantListener *Listener
+		h    Handler
+		addr netip.AddrPort
 	}{
-		"success": {addr: netip.MustParseAddrPort("127.0.0.1:7676"), h: func() Handler { m := &mock{}; return m.Handle }()},
+		"success": {addr: netip.MustParseAddrPort("127.0.0.1:7676"), h: &mock{}},
 	}
 	for name, tt := range tests {
 		t.Run(name, func(t *testing.T) {
-			s := &Listener{Addr: tt.addr}
-			t.Logf("before: %+v", s)
+			s, err := NewServer("lo", net.UDPAddrFromAddrPort(tt.addr), tt.h)
+			if err != nil {
+				t.Fatal(err)
+			}
 			ctx, done := context.WithCancel(context.Background())
 			defer done()
 
-			go s.ListenAndServe(ctx, tt.h)
+			go s.Serve(ctx)
 
 			// make client calls
 			d, err := dhcp(ctx)
@@ -103,78 +114,3 @@ func TestListenAndServe(t *testing.T) {
 		})
 	}
 }
-
-func TestListenerAndServe(t *testing.T) {
-	tests := map[string]struct {
-		h    Handler
-		addr netip.AddrPort
-		err  error
-	}{
-		"noop handler":             {h: func() Handler { m := &noop.Handler{}; return m.Handle }(), addr: netip.MustParseAddrPort("0.0.0.0:7678")},
-		"no handler":               {addr: netip.MustParseAddrPort("0.0.0.0:7678")},
-		"mock handler":             {h: func() Handler { m := &mock{}; return m.Handle }(), addr: netip.MustParseAddrPort("0.0.0.0:7678")},
-		"success use default addr": {h: func() Handler { m := &mock{}; return m.Handle }()},
-	}
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			s := &Listener{
-				Addr: tt.addr,
-			}
-			ctx, done := context.WithTimeout(context.Background(), time.Millisecond*100)
-			defer done()
-
-			err := s.ListenAndServe(ctx, tt.h)
-			if err != tt.err && err.Error() != "failed to create udp connection: cannot bind to port 67: permission denied" && !errors.Is(err, ErrNoConn) && !closeConnErr(err) { //nolint:errorlint // nil pointer dereference without this.
-				t.Log(err)
-				t.Fatalf("got: %T, wanted: %T or ErrNoConn", err, &net.OpError{})
-			}
-		})
-	}
-}
-
-func closeConnErr(err error) bool {
-	l, ok := err.(*net.OpError)
-	if !ok {
-		return false
-	}
-
-	return l.Op == "close"
-}
-
-func TestServe(t *testing.T) {
-	tests := map[string]struct {
-		h    Handler
-		addr netip.AddrPort
-		err  error
-	}{
-		"noop handler": {addr: netip.MustParseAddrPort("0.0.0.0:7676"), h: func() Handler { m := &noop.Handler{}; return m.Handle }()},
-		"no handler":   {addr: netip.MustParseAddrPort("0.0.0.0:7678")},
-	}
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			ctx, done := context.WithTimeout(context.Background(), time.Millisecond*100)
-			defer done()
-			go func() {
-				<-ctx.Done()
-			}()
-
-			err := Serve(ctx, nil, tt.h)
-			switch err.(type) {
-			case *net.OpError:
-			default:
-				if !errors.Is(err, ErrNoConn) {
-					t.Fatalf("got: %T, wanted: %T or ErrNoConn", err, &net.OpError{})
-				}
-			}
-		})
-	}
-}
-
-func TestNoConnError(t *testing.T) {
-	want := "no connection specified"
-	got := ErrNoConn
-	if diff := cmp.Diff(got.Error(), want); diff != "" {
-		t.Fatal(diff)
-	}
-}
-*/
