+++ example/kubeBackend/main.go
@@ -4,6 +4,7 @@ package main
 import (
 	"context"
 	"log"
+	"net"
 	"net/netip"
 	"net/url"
 	"os"
@@ -12,6 +13,7 @@ import (
 
 	"github.com/equinix-labs/otel-init-go/otelinit"
 	"github.com/go-logr/stdr"
+	"github.com/insomniacslk/dhcp/dhcpv4/server4"
 	"github.com/tinkerbell/dhcp"
 	"github.com/tinkerbell/dhcp/backend/kube"
 	"github.com/tinkerbell/dhcp/handler"
@@ -48,7 +50,7 @@ func main() {
 		OTELEnabled: true,
 		Backend:     backend,
 	}
-	conn, err := dhcp.NewConn(netip.MustParseAddrPort("0.0.0.0:67"))
+	conn, err := server4.NewIPv4UDPConn("", net.UDPAddrFromAddrPort(netip.MustParseAddrPort("0.0.0.0:67")))
 	if err != nil {
 		panic(err)
 	}
